working atomLegend
