+++ src/app/components/atomComponentVisual.jsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import * as d3 from 'd3';
 
-function AtomComponentVisual({ componentAtomTree, filteredSnapshot, selectedRecoilValue }) {
+function AtomComponentVisual({ componentAtomTree, filteredSnapshot, selectedRecoilValue, atoms, selectors }) {
 
   // set the heights and width of the tree to be passed into treeMap function
   const width = 600;
@@ -9,17 +9,16 @@ function AtomComponentVisual({ componentAtomTree, filteredSnapshot, selectedReco
   // this state allows the canvas to stay at the zoom level on multiple re-renders
   const [{ x, y, k }, setZoomState] = useState({ x: 0, y: 0, k: 0 });
 
-  const atoms = {};
-  const selectors = {};
-  if (filteredSnapshot) {
-    for (let [recoilValueName, object] of Object.entries(filteredSnapshot)) {
-      if (object.type === 'RecoilState') {
-        atoms[recoilValueName] = object.contents;
-      } else {
-        selectors[recoilValueName] = object.contents;
-      }
-    }
-  }
+    // create objects to hold the selected atom or selector and all of it's relationships to other atoms/selectors
+    // const selectorToAtom = {};
+    // const atomToSelector = {};
+    // if (selectedRecoilValue[1] === 'selector'){
+    //   selectorToAtom[selectedRecoilValue[0]] = filteredSnapshot[selectedRecoilValue[0]].nodeDeps;
+    // }
+    // if (selectedRecoilValue[1] === 'atom'){
+    //   atomToSelector[selectedRecoilValue[0]] = filteredSnapshot[selectedRecoilValue[0]].nodeToNodeSubscriptions;
+    // }
+
   useEffect(() => {
     setZoomState(d3.zoomTransform(d3.select('#canvas').node()));
   }, [componentAtomTree]);
@@ -170,9 +169,10 @@ function AtomComponentVisual({ componentAtomTree, filteredSnapshot, selectedReco
       function colorComponents(d) {
         // if component node cointains recoil atoms or selectors, make it orange red or yellow, otherwise keep node gray
         if (d.data.recoilNodes) {
-          if (d.data.recoilNodes.includes(selectedRecoilValue)){
+          if (d.data.recoilNodes.includes(selectedRecoilValue[0])){
             return 'white';
           }
+          
           let hasAtom = false;
           let hasSelector = false;
           for (let i = 0; i < d.data.recoilNodes.length; i++) {
